Separated type change events into separate directory and files
